+++ app/Werewolf/Command/Pass.hs
@@ -16,6 +16,7 @@ module Werewolf.Command.Pass (
 
 import Control.Monad.Except
 import Control.Monad.Extra
+import Control.Monad.Random
 import Control.Monad.State
 import Control.Monad.Writer
 
@@ -27,7 +28,7 @@ import Game.Werewolf.Command.Witch
 import Werewolf.Game
 import Werewolf.Messages
 
-handle :: MonadIO m => Text -> Text -> m ()
+handle :: (MonadIO m, MonadRandom m) => Text -> Text -> m ()
 handle callerName tag = do
     unlessM (doesGameExist tag) $ exitWith failure
         { messages = [noGameRunningMessage callerName]
@@ -37,6 +38,7 @@ handle callerName tag = do
 
     let command = passCommand callerName
 
-    case runExcept (runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game) of
+    result <- runExceptT . runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game
+    case result of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeOrDeleteGame tag game' >> exitWith success { messages = messages }