+++ app/Werewolf/Command/Vote.hs
@@ -20,6 +20,7 @@ module Werewolf.Command.Vote (
 import Control.Lens
 import Control.Monad.Except
 import Control.Monad.Extra
+import Control.Monad.Random
 import Control.Monad.State
 import Control.Monad.Writer
 
@@ -36,7 +37,7 @@ data Options = Options
     { argTarget :: Text
     } deriving (Eq, Show)
 
-handle :: MonadIO m => Text -> Text -> Options -> m ()
+handle :: (MonadIO m, MonadRandom m) => Text -> Text -> Options -> m ()
 handle callerName tag (Options targetName) = do
     unlessM (doesGameExist tag) $ exitWith failure
         { messages = [noGameRunningMessage callerName]
@@ -51,6 +52,7 @@ handle callerName tag (Options targetName) = do
                 { messages = [playerCannotDoThatRightNowMessage callerName]
                 }
 
-    case runExcept (runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game) of
+    result <- runExceptT . runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game
+    case result of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeOrDeleteGame tag game' >> exitWith success { messages = messages }